* Moved JPA from specific/ to extensions/ * standardized extensions artifact naming to radargun-*
