+++ extensions/example-extension/src/main/java/org/radargun/stages/ExampleStage.java
