+++ extensions/jpa/pom.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
    <parent>
-      <artifactId>radargun-parent</artifactId>
       <groupId>org.radargun</groupId>
+      <artifactId>radargun-extensions</artifactId>
       <version>3.0.0-SNAPSHOT</version>
-      <relativePath>../../parent/pom.xml</relativePath>
+      <relativePath>../pom.xml</relativePath>
    </parent>
    <modelVersion>4.0.0</modelVersion>
 
@@ -18,10 +18,11 @@
          <artifactId>radargun-core</artifactId>
          <version>${project.version}</version>
       </dependency>
+      <!-- Let's use JPA 2.1, though there is no maven spec-release -->
       <dependency>
-         <groupId>javax.persistence</groupId>
-         <artifactId>persistence-api</artifactId>
-         <version>1.0.2</version>
+         <groupId>org.hibernate.javax.persistence</groupId>
+         <artifactId>hibernate-jpa-2.1-api</artifactId>
+         <version>1.0.0.Final</version>
          <scope>provided</scope>
       </dependency>
    </dependencies>
