+++ extensions/jpa/src/main/java/org/radargun/jpa/entities/BasicEntity.java
@@ -1,4 +1,4 @@
-package org.radargun.jpa;
+package org.radargun.jpa.entities;
 
 import org.radargun.logging.Log;
 import org.radargun.logging.LogFactory;
