+++ extensions/jpa/src/main/java/org/radargun/jpa/entities/EmbeddedContentEntity.java
@@ -1,4 +1,4 @@
-package org.radargun.jpa;
+package org.radargun.jpa.entities;
 
 import org.radargun.stages.cache.generators.JpaValueGenerator;
